[xharness] Don't try killing a process that has already exited.

+++ tests/xharness/DeviceLogCapturer.cs
@@ -55,6 +55,9 @@ namespace xharness
 
 		public void StopCapture ()
 		{
+			if (process.HasExited)
+				return;
+			
 			process.Kill ();
 			if (!streamEnds.Wait (TimeSpan.FromSeconds (5))) {
 				Harness.Log ("Could not kill 'mtouch --logdev' process in 5 seconds.");
